pyproject.toml
[tool.poetry]
name = "wsdm"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kaggle = "^1.6.17"
jinja2 = "^3.1.5"
ipykernel = "^6.29.5"
numpy = "^2.2.2"
pandas = "^2.2.3"
pyarrow = "^19.0.0"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
transformers = "^4.48.3"
torch = "^2.6.0"
accelerate = "^1.3.0"
logzero = "^1.7.0"
datasets = "^3.2.0"
openai = "^1.63.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
